Fixed screen tearing issue in PPU
The visual effect of this is a distorted line that moves up or down the
screen.

The root cause is that we're emulating the Gameboy for a _full_
frame (including VBLANK!), and then rendering the result to the screen.
But, the CPU runs for a couple of cycles at the end of VBLANK, which
may push the PPU into the next frame. We won't be drawing anything to
the next frame, but we will end up skewing its mode timings!

To fix this, I ended up:

1. Marking the frame buffer as "ready" as soon as we enter VBLANK
2. Moving the frame loop into the emulator (i.e., outside of lib)

With (2), as soon as the frame buffer is ready, we can pass it onto
SDL for rendering to the texture.