Run frame from VBLANK to VBLANK
To avoid relying on cycles per frame, which is not going to work with
CPU speed switching, we instead run the Gameboy until the next frame
is ready. Once we get a frame, we render it in SDL, and move on to the
next one. The next VBLANK will be executed during the next frame loop
iteration.

In addition, I switched to `spin_sleep` to get slightly more accurate
frame sleep times, especially on Windows.