Basic save state support
This uses Serde + Bincode to serialize all relevant state to a
single file on disk. The feature is enabled by default in the lib.

The state file includes:

- CPU state
- PPU state and frame buffer contents
- Work RAM
- I/O registers
- DMA state
- Timer state
- Joypad state

The ROM content is _not_ serialized to disk. Instead, the ROM is
loaded into memory from the provided ROM file, just like we do on
regular init. ROM state is saved, however (i.e., active bank).

As an example, for Donkey Kong Country, the size of the save state file
is around 120 KB.