Removed all file-based persistence logic in library
With these changes, both save states and persistent state (RAM, RTC)
are exposed by the library as raw bytes.

For save states, the emulator can take care of how to save the state for
later use. For persistent ROM state, I've added a few methods that
allow the upper layer (emulator) to persist this state at the end of
each frame and load it on init. This is all the responsibility of the
emulator, though.