Split up RAM and RTC persistence
For RAM, we now only return the RAM contents if the RAM has been written
to since the last frame. This allows us to avoid writing out the entire
RAM on every frame.

To achieve this, I had to revert back to splitting the RAM and RTC
state into separate files on disk.

Closes #8.