Minor changes to comments in PPU
P.S. This was a "failed" attempt to optimize the BG fetch method. The
idea was to replace the tile data fetch loop into a copy_from_slice
(memcpy) call. After looking at the generated ASM, it runs out that both
approaches are unrolled into a sequence of 16 movs, so... :)